refactor: [BE] support error 위치 변경
