+++ core/core-api/src/main/java/com/team/voteland/core/api/config/AsyncExceptionHandler.java
@@ -1,7 +1,7 @@
 package com.team.voteland.core.api.config;
 
-import com.team.voteland.core.support.error.CoreException;
 
+import com.team.voteland.support.error.CoreException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
