+++ core/core-api/src/main/java/com/team/voteland/core/support/response/ApiResponse.java
@@ -1,7 +1,8 @@
 package com.team.voteland.core.support.response;
 
-import com.team.voteland.core.support.error.ErrorMessage;
-import com.team.voteland.core.support.error.ErrorType;
+
+import com.team.voteland.support.error.ErrorMessage;
+import com.team.voteland.support.error.ErrorType;
 
 public class ApiResponse<S> {
