+++ core/domain-user/build.gradle
@@ -6,6 +6,7 @@ dependencies {
     implementation project(':storage:db-core')
     implementation project(':support:security')
     implementation project(':support:event')
+    implementation project(':support:error')
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
