+++ core/domain-vote/build.gradle
@@ -4,8 +4,10 @@ jar.enabled = true
 dependencies {
     implementation project(':core:core-enum')
     implementation project(':storage:db-core')
+    implementation project(':support:error')
     implementation project(':support:event')
     implementation project(':core:domain-user')
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
 }
