+++ support/error/src/main/java/com/team/voteland/support/error/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.team.voteland.core.support.error;
+package com.team.voteland.support.error;
 
 public class ErrorMessage {
